handle dropdown picker view
