+++ Components/BaseComponent/DemoComponent/DemoComponent/InputFieldViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 class InputFieldViewController: UIViewController {
     
-    @IBOutlet weak var inputField: BaseInputField!
+    @IBOutlet weak var inputField: BaseDropDown!
     @IBOutlet weak var inputField2: BaseInputField!
     @IBOutlet weak var inputField3: BaseInputField!
     @IBOutlet weak var inputField4: BaseInputField!
